+++ signify/authenticode/cli.py
@@ -1,18 +1,18 @@
 from __future__ import annotations
 
+import argparse
 import logging
-import re
 import pathlib
-import sys
+import re
 import textwrap
-from typing import Any
+from typing import Any, cast
 
 from signify.authenticode import AuthenticodeFile, CertificateTrustList
-from signify.authenticode.indirect_data import IndirectData
+from signify.authenticode.indirect_data import IndirectData, PeImageData, SigInfo
 from signify.authenticode.signed_data import AuthenticodeSignature
 from signify.authenticode.signer_info import AuthenticodeSignerInfo
 from signify.authenticode.tsp import RFC3161SignedData
-from signify.pkcs7 import SignerInfo, SignedData
+from signify.pkcs7 import SignedData, SignerInfo
 from signify.x509 import Certificate
 
 
@@ -120,7 +120,9 @@ def describe_signer_info(signer_info: SignerInfo) -> list[str]:
             result += [
                 "Countersigner (nested RFC3161):",
                 indent_text(
-                    *describe_signed_data(signer_info.countersigner),
+                    *describe_signed_data(
+                        cast(RFC3161SignedData, signer_info.countersigner)
+                    ),
                     indent=4,
                 ),
             ]
@@ -134,7 +136,9 @@ def describe_indirect_data(indirect_data: IndirectData) -> list[str]:
         f"Digest: {indirect_data.digest.hex()}",
         f"Content type: {indirect_data.content_type}",
     ]
-    if indirect_data.content_type == "microsoft_spc_pe_image_data":
+    if indirect_data.content_type == "microsoft_spc_pe_image_data" and isinstance(
+        indirect_data.content, PeImageData
+    ):
         pe_image_data = indirect_data.content
         result += [
             "",
@@ -158,7 +162,9 @@ def describe_indirect_data(indirect_data: IndirectData) -> list[str]:
                 )
             ]
 
-    if indirect_data.content_type == "microsoft_spc_siginfo":
+    if indirect_data.content_type == "microsoft_spc_siginfo" and isinstance(
+        indirect_data.content, SigInfo
+    ):
         siginfo = indirect_data.content
         result += [
             "",
@@ -172,17 +178,20 @@ def describe_indirect_data(indirect_data: IndirectData) -> list[str]:
     return result
 
 
-def describe_signed_data(signed_data: SignedData):
-    result = [
-        "Included certificates:",
-        *[format_certificate(cert) for cert in signed_data.certificates],
-        "",
-        "Signer:",
-        indent_text(*describe_signer_info(signed_data.signer_info), indent=4),
-        "",
-        f"Digest algorithm: {signed_data.digest_algorithm.__name__}",
-        f"Content type: {signed_data.content_type}",
-    ]
+def describe_signed_data(signed_data: SignedData) -> list[str]:
+    result = []
+
+    if len(signed_data.signer_infos) == 1:
+        result += [
+            "Included certificates:",
+            *[format_certificate(cert) for cert in signed_data.certificates],
+            "",
+            "Signer:",
+            indent_text(*describe_signer_info(signed_data.signer_info), indent=4),
+            "",
+            f"Digest algorithm: {signed_data.digest_algorithm.__name__}",
+            f"Content type: {signed_data.content_type}",
+        ]
 
     if isinstance(signed_data, AuthenticodeSignature) and signed_data.indirect_data:
         result += [
@@ -250,28 +259,63 @@ def describe_signed_data(signed_data: SignedData):
     return result
 
 
-def main(*filenames: str):
-    logging.basicConfig(level=logging.DEBUG)
+def main() -> None:
+    parser = argparse.ArgumentParser(description="Verify Authenticode signature")
+    parser.add_argument(
+        "filenames", nargs="*", help="Filenames to verify", type=pathlib.Path
+    )
+    parser.add_argument(
+        "--catalog",
+        action="append",
+        help="Catalog files to use in verification",
+        type=pathlib.Path,
+    )
+    parser.add_argument(
+        "--verbose",
+        "-v",
+        action="store_true",
+        help="Provide full details on the signatures.",
+    )
+    args = parser.parse_args()
 
-    for filename in filenames:
+    if args.verbose:
+        logging.basicConfig(level=logging.DEBUG)
+
+    for filename in args.filenames:
         print(f"{filename}:")
-        with pathlib.Path(filename).open("rb") as file_obj:
-            try:
-                signed_file = AuthenticodeFile.from_stream(file_obj)
-                for signed_data in signed_file.signatures:
-                    print(indent_text(*describe_signed_data(signed_data), indent=4))
-                    print("--------")
+        try:
+            with filename.open("rb") as file_obj:
+                signed_file = AuthenticodeFile.from_stream(file_obj, allow_flat=True)
 
+                # Add catalogs
+                for catalog in args.catalog or ():
+                    with catalog.open("rb") as catalog_obj:
+                        signed_file.add_catalog(catalog_obj)
+
+                # When verbose, print the full signatures
+                if args.verbose:
+                    for signed_data in signed_file.signatures:
+                        print(indent_text(*describe_signed_data(signed_data), indent=4))
+                        print("--------")
+
+                # Get the result
                 result, e = signed_file.explain_verify()
-                print(result)
-                if e:
-                    raise e
-                    print(repr(e))
 
-            except Exception as e:
-                raise
+                # When verbose, print the full result, or print the summary when not
+                # verbose.
+                if args.verbose:
+                    print(result)
+                    if e:
+                        print(repr(e))
+                else:
+                    print(f"    Status: {result.name}")
+
+        except Exception as e:
+            if args.verbose:
                 print("    Error while parsing: " + str(e))
+            else:
+                print("    Status: ERROR")
 
 
 if __name__ == "__main__":
-    main(*sys.argv[1:])
+    main()